requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
telethon==1.33.0
aiohttp==3.8.5
requests==2.31.0
